IMPORTANT!!!!!! Changes to login & a note on session start
Okay so read dis plz, basically check the changes in signIn.php, what I've done is that I've created an array called $row whose job is to collect ALL the user's data and throw it into the session variable, that's how we get access to displayname.

What you had done before was simply put the data that was entered by $_POST into the session variable and that's like...not from the database, I hope you understand what I mean.

SESSIONS MUST BE STARTED AT THE BEGINNING OF THE PAGE!
you had a session_start() called on line 27, that's unforgwivabwe, you can only start sessions in the beginning, so that the entire page gets access to the session variables.

Now we have the displayname showing in everyone's profiles.